+++ dexterous_developer_internal/src/bevy_support/mod.rs
@@ -10,7 +10,7 @@ mod hot_internal;
 pub use types::*;
 
 #[cfg(feature = "hot_internal")]
-pub use hot_internal::*;
+pub use hot_internal::{ReloadableAppContents, HotReloadPlugin};
 
 #[cfg(not(feature = "hot_internal"))]
 pub use cold::*;